Adds many pieces to the front-end
I messed up the initial fork I created to work on front-end stuff and with it, all the commits Drew and I had made to it. So I will try to recap the changes made in this commit: First we established our initial landing page and styled out the correct areas of our webpage using CSS. Then, Brandon worked on DraftBoard.js which includes a Start Draft button that initiates a timer and alerts the user who is currently drafting. After all teams have presumably drafted, a check is made and the draft timer ends. Below that, there are 10 team place holders, along with their 10 respective player slots in a grid format. Drew worked on AvailablePlayerList.js and he was able to fetch the player data from our MongoDB and display it on his webpage, we still need to incorporate this into App.js.